requirements.txt
geoserver-rest
ipython
openpyxl
pandas
python-dotenv